fix(core): refactor to use form control mixins
