+++ projects/core/src/button/button.test.lighthouse.ts
@@ -18,6 +18,6 @@ describe('button lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(13.6);
+    expect(report.payload.javascript.kb).toBeLessThan(14.2);
   });
 });
