+++ projects/core/src/chat-message/chat-message.test.lighthouse.ts
@@ -16,6 +16,6 @@ describe('button lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(13.7);
+    expect(report.payload.javascript.kb).toBeLessThan(14.2);
   });
 });
