+++ projects/core/src/dialog/dialog.test.lighthouse.ts
@@ -25,6 +25,6 @@ describe('dialog lighthouse report', () => {
     expect(report.scores.performance).toBeGreaterThan(97); // bfcache
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(24.5);
+    expect(report.payload.javascript.kb).toBeLessThan(24.6);
   });
 });
