+++ projects/core/src/drawer/drawer.test.lighthouse.ts
@@ -18,6 +18,6 @@ describe('drawer lighthouse report', () => {
     expect(report.scores.performance).toBeGreaterThan(98); // bfcache
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(25);
+    expect(report.payload.javascript.kb).toBeLessThan(25.1);
   });
 });
