+++ projects/core/src/dropzone/dropzone.test.lighthouse.ts
@@ -16,6 +16,6 @@ describe('dropzone lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(18.1);
+    expect(report.payload.javascript.kb).toBeLessThan(18.9);
   });
 });
