+++ projects/core/src/internal/controllers/type-command.controller.test.ts
@@ -8,14 +8,24 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { createFixture, removeFixture, untilEvent, emulateClick, elementIsStable } from '@internals/testing';
 import { TypeCommandController } from '@nvidia-elements/core/internal';
 
-@customElement('type-command-controller-test-element')
-class TypeCommandControllerTestElement extends LitElement {
+class TypeCommandControllerTestElementBase extends LitElement {
   @property({ type: String }) command: string;
   @property({ type: String, attribute: 'commandfor' }) commandfor: string;
   @property({ type: Object }) commandForElement: HTMLElement;
   @property({ type: Boolean, attribute: 'readonly' }) readOnly = false;
   @property({ type: Boolean }) disabled: boolean;
+}
+
+@customElement('type-command-controller-test-element')
+class TypeCommandControllerTestElement extends TypeCommandControllerTestElementBase {
   _typeCommandController = new TypeCommandController<TypeCommandControllerTestElement>(this);
+}
+
+@customElement('manual-type-command-controller-test-element')
+class ManualTypeCommandControllerTestElement extends TypeCommandControllerTestElementBase {
+  _typeCommandController = new TypeCommandController<ManualTypeCommandControllerTestElement>(this, {
+    trigger: 'manual'
+  });
 }
 
 describe('type-command.controller', () => {
@@ -46,6 +56,10 @@ describe('type-command.controller', () => {
       expect(source).toBe(element);
       expect(command).toBe('--test');
     });
+
+    it('should expose the resolved target', () => {
+      expect(element._typeCommandController.target).toBe(target);
+    });
   });
 
   describe('commandForElement property', () => {
@@ -69,6 +83,73 @@ describe('type-command.controller', () => {
       expect(source).toBe(element);
       expect(command).toBe('--test');
     });
+
+    it('should expose the commandForElement target', () => {
+      expect(element._typeCommandController.target).toBe(target);
+    });
+
+    it('should prefer commandForElement over commandfor', async () => {
+      removeFixture(fixture);
+      fixture = await createFixture(
+        html`
+          <type-command-controller-test-element command="--test" commandfor="target-id"></type-command-controller-test-element>
+          <div id="target-id"></div>
+          <div id="target-property"></div>
+        `
+      );
+      element = fixture.querySelector<TypeCommandControllerTestElement>('type-command-controller-test-element');
+      target = fixture.querySelector<HTMLElement>('#target-property');
+      const targetById = fixture.querySelector<HTMLElement>('#target-id');
+      element.commandForElement = target;
+      await elementIsStable(element);
+
+      let commandForIdFired = false;
+      targetById.addEventListener('command', () => (commandForIdFired = true));
+      const event = untilEvent<Event & { source: HTMLElement; command: string }>(target, 'command');
+      await emulateClick(element);
+
+      expect((await event).command).toBe('--test');
+      expect(commandForIdFired).toBe(false);
+    });
+  });
+
+  describe('manual trigger', () => {
+    let manualElement: ManualTypeCommandControllerTestElement;
+
+    beforeEach(async () => {
+      fixture = await createFixture(
+        html`
+          <manual-type-command-controller-test-element
+            command="--test"
+            commandfor="target"
+          ></manual-type-command-controller-test-element>
+          <div id="target"></div>
+        `
+      );
+      manualElement = fixture.querySelector<ManualTypeCommandControllerTestElement>(
+        'manual-type-command-controller-test-element'
+      );
+      target = fixture.querySelector<HTMLElement>('#target');
+      await elementIsStable(manualElement);
+    });
+
+    it('should not trigger command event when clicked', async () => {
+      let fired = false;
+      target.addEventListener('command', () => (fired = true));
+      await emulateClick(manualElement);
+
+      expect(fired).toBe(false);
+    });
+
+    it('should dispatch command event explicitly', async () => {
+      const event = untilEvent<Event & { source: HTMLElement; command: string }>(target, 'command');
+      const dispatched = manualElement._typeCommandController.dispatchCommand();
+
+      expect(dispatched).toBe(true);
+      const { source, command } = await event;
+      expect(source).toBe(manualElement);
+      expect(command).toBe('--test');
+    });
   });
 
   describe('readonly', () => {
@@ -88,6 +169,28 @@ describe('type-command.controller', () => {
       let fired = false;
       target.addEventListener('command', () => (fired = true));
       await emulateClick(element);
+      expect(fired).toBe(false);
+    });
+
+    it('should not dispatch command event when deprecated readonly property is true', async () => {
+      removeFixture(fixture);
+      fixture = await createFixture(
+        html`
+          <type-command-controller-test-element command="--test" commandfor="target"></type-command-controller-test-element>
+          <div id="target"></div>
+        `
+      );
+      element = fixture.querySelector<TypeCommandControllerTestElement>('type-command-controller-test-element');
+      target = fixture.querySelector<HTMLElement>('#target');
+      (element as TypeCommandControllerTestElement & { readonly: boolean }).readonly = true;
+      await elementIsStable(element);
+
+      let fired = false;
+      target.addEventListener('command', () => (fired = true));
+
+      await emulateClick(element);
+
+      expect(element._typeCommandController.dispatchCommand()).toBe(false);
       expect(fired).toBe(false);
     });
   });
