+++ projects/core/src/internal/index.ts
@@ -4,7 +4,6 @@
 /**
  * @internal private common utilities for Elements
  */
-export * from './base/button.js';
 export * from './controllers/audit.controller.js';
 export * from './controllers/i18n.controller.js';
 export * from './controllers/keynav-grid.controller.js';
