+++ projects/core/src/preferences-input/preferences-input.test.lighthouse.ts
@@ -16,7 +16,6 @@ describe('preferences-input lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(31.5);
+    expect(report.payload.javascript.kb).toBeLessThan(32.1);
   });
 });
-
