+++ projects/core/src/resize-handle/resize-handle.test.lighthouse.ts
@@ -16,6 +16,6 @@ describe('resize-handle lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(12.6);
+    expect(report.payload.javascript.kb).toBeLessThan(13.3);
   });
 });
