+++ projects/core/src/sort-button/sort-button.test.lighthouse.ts
@@ -16,6 +16,6 @@ describe('sort-button lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(18.5);
+    expect(report.payload.javascript.kb).toBeLessThan(18.6);
   });
 });
