+++ projects/core/src/steps/steps.test.lighthouse.ts
@@ -20,6 +20,6 @@ describe('steps lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(23.2);
+    expect(report.payload.javascript.kb).toBeLessThan(23.6);
   });
 });
