+++ projects/core/src/tag/tag.test.lighthouse.ts
@@ -16,6 +16,6 @@ describe('tag lighthouse report', () => {
     expect(report.scores.performance).toBe(100);
     expect(report.scores.accessibility).toBe(100);
     expect(report.scores.bestPractices).toBe(100);
-    expect(report.payload.javascript.kb).toBeLessThan(18);
+    expect(report.payload.javascript.kb).toBeLessThan(18.4);
   });
 });
