+++ projects/core/src/tag/tag.ts
@@ -1,13 +1,13 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { html } from 'lit';
+import { html, LitElement } from 'lit';
 import { property } from 'lit/decorators/property.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { ButtonFormControlMixin } from '@nvidia-elements/forms/mixins';
 import type { Color, Prominence } from '@nvidia-elements/core/internal';
 import {
   I18nController,
-  BaseButton,
   TypeClosableController,
   useStyles,
   colorStateStyles,
@@ -39,7 +39,7 @@ import styles from './tag.css?inline';
  * @aria https://www.w3.org/WAI/ARIA/apg/patterns/button/
  */
 @scopedRegistry()
-export class Tag extends BaseButton {
+export class Tag extends ButtonFormControlMixin(LitElement) {
   static styles = useStyles([colorStateStyles, styles]);
 
   static readonly metadata = {
